refactor: admin 관련 api admin.ts로 통일
