+++ src/hooks/useLoanApplicationsQuery.ts
@@ -3,7 +3,7 @@ import React, { useMemo } from 'react';
 import { useQuery } from '@tanstack/react-query';
 import dayjs from 'dayjs';
 
-import { getLoanApplications } from '@/apis/AdminLoans';
+import { getLoanApplications } from '@/apis/admin';
 import { LoanFilterType } from '@/types/loan.filter.type';
 import { filtersToLoanApplicationParams } from '@/utils/loanApplicationParams';
 import { RawLoanApplication } from '@/types/admin.type';
