+++ src/hooks/useMembersQuery.ts
@@ -3,7 +3,7 @@ import React, { useMemo } from 'react';
 import { useQuery } from '@tanstack/react-query';
 import dayjs from 'dayjs';
 
-import { getMembers } from '@/apis/adminMembers';
+import { getMembers } from '@/apis/admin';
 import { FilterType } from '@/types/filter.type';
 import { filtersToParams } from '@/utils/memberParams';
 import { RawMember } from '@/types/admin.type';
